set correct flags to init geometry, tethex interface update, print tria upd, minor fixes

+++ source/river/riversim.hpp
@@ -68,8 +68,8 @@ namespace River
     int numOfRegions = 1;
     vector<double> regions = {};
     //out only
-    vector<double> edges = {};
-    vector<double> edgeMarkers = {};
+    vector<int> edges = {};
+    vector<int> edgeMarkers = {};
   };
